<==== Mercurial data ====>
Branch: master

+++ mediaserver/src/recorder/storage_manager.cpp
@@ -46,6 +46,47 @@ private:
     QnStorageManager* m_storageManager;
 };
 
+class ScanMediaFilesTask: public QnLongRunnable
+{
+private:
+    QMap<DeviceFileCatalogPtr, qint64> m_data; // key - catalog, value - start scan time;
+    QnStorageManager* m_owner;
+    QnStorageResourceList m_storagesToScan;
+    qint64 m_endScanTime;
+public:
+    ScanMediaFilesTask(QnStorageManager* owner): QnLongRunnable(), m_owner(owner)
+    {
+        foreach(const DeviceFileCatalogPtr& catalog, owner->m_devFileCatalog[QnServer::LowQualityCatalog])
+            m_data.insert(catalog, catalog->lastChunkStartTime());
+        foreach(const DeviceFileCatalogPtr& catalog, owner->m_devFileCatalog[QnServer::HiQualityCatalog])
+            m_data.insert(catalog, catalog->lastChunkStartTime());
+        m_storagesToScan = m_owner->getStorages();
+        m_endScanTime = qnSyncTime->currentMSecsSinceEpoch();
+
+        m_owner->m_rebuildProgress = 0;
+        m_owner->m_rebuildState = QnStorageManager::RebuildState_Initial;
+    }
+    virtual void run() override
+    {
+        int steps = m_data.size() * m_storagesToScan.size();
+        foreach(const QnStorageResourcePtr& storage, m_storagesToScan)
+        {
+            for(auto itr = m_data.begin(); itr != m_data.end(); ++itr) 
+            {
+                DeviceFileCatalog::ScanFilter filter;
+                filter.scanPeriod.startTimeMs = itr.value();
+                filter.scanPeriod.durationMs = qMax(1ll, m_endScanTime - filter.scanPeriod.startTimeMs);
+                m_owner->partialMediaScan(itr.key(), storage, filter);
+                m_owner->m_rebuildProgress += 1.0 / steps;
+                if (needToStop())
+                    return;
+            }
+        }
+        
+        m_owner->m_rebuildProgress = 1.0;
+        m_owner->m_rebuildState = QnStorageManager::RebuildState_None;
+    }
+};
 
 class TestStorageThread: public QnLongRunnable
 {
@@ -86,6 +127,7 @@ QnStorageManager::QnStorageManager():
     m_rebuildState(RebuildState_None),
     m_rebuildProgress(0),
     m_asyncRebuildTask(0),
+    m_asyncPartialScan(0),
     m_initInProgress(true)
 {
     m_lastTestTime.restart();
@@ -97,17 +139,14 @@ QnStorageManager::QnStorageManager():
     connect(qnResPool, &QnResourcePool::resourceAdded, this, &QnStorageManager::onNewResource, Qt::DirectConnection);
 }
 
-std::deque<DeviceFileCatalog::Chunk> QnStorageManager::correctChunksFromMediaData(const DeviceFileCatalogPtr &fileCatalog, const QnStorageResourcePtr &storage, const std::deque<DeviceFileCatalog::Chunk>& chunks)
+//std::deque<DeviceFileCatalog::Chunk> QnStorageManager::correctChunksFromMediaData(const DeviceFileCatalogPtr &fileCatalog, const QnStorageResourcePtr &storage, const std::deque<DeviceFileCatalog::Chunk>& chunks)
+void QnStorageManager::partialMediaScan(const DeviceFileCatalogPtr &fileCatalog, const QnStorageResourcePtr &storage, const DeviceFileCatalog::ScanFilter& filter)
 {
     QnServer::ChunksCatalog catalog = fileCatalog->getCatalog();
-
+    
     /* Check new records, absent in the DB */
     QVector<DeviceFileCatalog::EmptyFileInfo> emptyFileList;
     QString rootDir = fileCatalog->rootFolder(storage, catalog);
-    DeviceFileCatalog::ScanFilter filter;
-    if (!chunks.empty())
-        filter.scanAfter = chunks[chunks.size()-1];
-
 
     QMap<qint64, DeviceFileCatalog::Chunk> newChunksMap;
     fileCatalog->scanMediaFiles(rootDir, storage, newChunksMap, emptyFileList, filter);
@@ -119,14 +158,20 @@ std::deque<DeviceFileCatalog::Chunk> QnStorageManager::correctChunksFromMediaDat
         qnFileDeletor->deleteFile(emptyFile.fileName);
 
     // add to DB
-    QnStorageDbPtr sdb = m_chunksDB[storage->getPath()];
+    QnStorageDbPtr sdb = getSDB(storage);
     QString cameraUniqueId = fileCatalog->cameraUniqueId();
-    foreach(const DeviceFileCatalog::Chunk& chunk, newChunks)
-        sdb->addRecord(cameraUniqueId, catalog, chunk);
-    sdb->flushRecords();
+    foreach(const DeviceFileCatalog::Chunk& chunk, newChunks) {
+        if (QnResource::isStopping())
+            break;
+        if (sdb)
+            sdb->addRecord(cameraUniqueId, catalog, chunk);
+    }
+    if (sdb)
+        sdb->flushRecords();
     // merge chunks
-    return DeviceFileCatalog::mergeChunks(chunks, newChunks);
+    fileCatalog->addChunks(newChunks);
 }
+
 
 void QnStorageManager::initDone()
 {
@@ -136,6 +181,9 @@ void QnStorageManager::initDone()
     disconnect(qnResPool);
     connect(qnResPool, &QnResourcePool::resourceAdded, this, &QnStorageManager::onNewResource, Qt::QueuedConnection);
     connect(qnResPool, &QnResourcePool::resourceRemoved, this, &QnStorageManager::onDelResource, Qt::QueuedConnection);
+
+    m_asyncPartialScan = new ScanMediaFilesTask(this);
+    m_asyncPartialScan->start();
 }
 
 QMap<QString, QSet<int>> QnStorageManager::deserializeStorageFile()
@@ -168,24 +216,9 @@ QMap<QString, QSet<int>> QnStorageManager::deserializeStorageFile()
 
 bool QnStorageManager::loadFullFileCatalog(const QnStorageResourcePtr &storage, bool isRebuild, qreal progressCoeff)
 {
-    QString simplifiedGUID = qnCommon->moduleGUID().toString();
-    simplifiedGUID = simplifiedGUID.replace("{", "");
-    simplifiedGUID = simplifiedGUID.replace("}", "");
-    QString fileName = closeDirPath(storage->getPath()) + QString::fromLatin1("%1_media.sqlite").arg(simplifiedGUID);
-    QString oldFileName = closeDirPath(storage->getPath()) + QString::fromLatin1("media.sqlite");
-    if (QFile::exists(oldFileName) && !QFile::exists(fileName))
-        QFile::rename(oldFileName, fileName);
-
-    QnStorageDbPtr sdb = m_chunksDB[storage->getPath()];
-    if (!sdb) {
-        sdb = m_chunksDB[storage->getPath()] = QnStorageDbPtr(new QnStorageDb(storage->getIndex()));
-        if (!sdb->open(fileName))
-        {
-            qWarning() << "can't initialize sqlLite database! Actions log is not created!";
-            return false;
-        }
-    }
-
+    QnStorageDbPtr sdb = getSDB(storage);
+    if (!sdb)
+        return false;
     if (!isRebuild)
     {
         if (!m_initInProgress)
@@ -205,13 +238,16 @@ bool QnStorageManager::loadFullFileCatalog(const QnStorageResourcePtr &storage, 
 
 void QnStorageManager::addDataFromDatabase(const QnStorageResourcePtr &storage)
 {
-    QnStorageDbPtr sdb = m_chunksDB[storage->getPath()];
-
+    QnStorageDbPtr sdb = getSDB(storage);
+    if (!sdb)
+        return;
+    
     // load from database
     foreach(const DeviceFileCatalogPtr& c, sdb->loadFullFileCatalog())
     {
         DeviceFileCatalogPtr fileCatalog = getFileCatalogInternal(c->cameraUniqueId(), c->getCatalog());
-        fileCatalog->addChunks(correctChunksFromMediaData(fileCatalog, storage, c->m_chunks));
+        fileCatalog->addChunks(c->m_chunks);
+        //fileCatalog->addChunks(correctChunksFromMediaData(fileCatalog, storage, c->m_chunks));
     }
 }
 
@@ -303,7 +339,7 @@ void QnStorageManager::loadFullFileCatalogFromMedia(const QnStorageResourcePtr &
 
         QString cameraUniqueId = fi.fileName();
 
-        qint64 rebuildEndTime = qnSyncTime->currentMSecsSinceEpoch() - 100 * 1000;
+        qint64 rebuildEndTime = qnSyncTime->currentMSecsSinceEpoch() - QnRecordingManager::RECORDING_CHUNK_LEN * 1250;
         DeviceFileCatalogPtr newCatalog(new DeviceFileCatalog(cameraUniqueId, catalog));
         QnTimePeriod rebuildPeriod = QnTimePeriod(0, rebuildEndTime);
         newCatalog->doRebuildArchive(storage, rebuildPeriod);
@@ -356,29 +392,55 @@ QSet<int> QnStorageManager::getDeprecateIndexList(const QString& p)
     return result;
 }
 
+QnStorageDbPtr QnStorageManager::getSDB(const QnStorageResourcePtr &storage)
+{
+    QMutexLocker lock(&m_sdbMutex);
+    QnStorageDbPtr sdb = m_chunksDB[storage->getPath()];
+    if (!sdb) 
+    {
+        QString simplifiedGUID = qnCommon->moduleGUID().toString();
+        simplifiedGUID = simplifiedGUID.replace("{", "");
+        simplifiedGUID = simplifiedGUID.replace("}", "");
+        QString fileName = closeDirPath(storage->getPath()) + QString::fromLatin1("%1_media.sqlite").arg(simplifiedGUID);
+        QString oldFileName = closeDirPath(storage->getPath()) + QString::fromLatin1("media.sqlite");
+        if (QFile::exists(oldFileName) && !QFile::exists(fileName))
+            QFile::rename(oldFileName, fileName);
+
+        sdb = m_chunksDB[storage->getPath()] = QnStorageDbPtr(new QnStorageDb(storage->getIndex()));
+        if (!sdb->open(fileName))
+        {
+            qWarning() << "can't initialize sqlLite database! Actions log is not created!";
+            return QnStorageDbPtr();
+        }
+    }
+    return sdb;
+}
+
 void QnStorageManager::addStorage(const QnStorageResourcePtr &storage)
 {
-    storage->setIndex(detectStorageIndex(storage->getPath()));
-    QMutexLocker lock(&m_mutexStorages);
-    m_storagesStatisticsReady = false;
+    {
+        storage->setIndex(detectStorageIndex(storage->getPath()));
+        QMutexLocker lock(&m_mutexStorages);
+        m_storagesStatisticsReady = false;
     
-    NX_LOG(QString("Adding storage. Path: %1. SpaceLimit: %2MiB. Currently available: %3MiB").arg(storage->getPath()).arg(storage->getSpaceLimit() / 1024 / 1024).arg(storage->getFreeSpace() / 1024 / 1024), cl_logINFO);
+        NX_LOG(QString("Adding storage. Path: %1. SpaceLimit: %2MiB. Currently available: %3MiB").arg(storage->getPath()).arg(storage->getSpaceLimit() / 1024 / 1024).arg(storage->getFreeSpace() / 1024 / 1024), cl_logINFO);
 
-    removeStorage(storage); // remove existing storage record if exists
-    //QnStorageResourcePtr oldStorage = removeStorage(storage); // remove existing storage record if exists
-    //if (oldStorage)
-    //    storage->addWritedSpace(oldStorage->getWritedSpace());
-    m_storageRoots.insert(storage->getIndex(), storage);
-    if (storage->isStorageAvailable())
-        storage->setStatus(Qn::Online);
+        removeStorage(storage); // remove existing storage record if exists
+        //QnStorageResourcePtr oldStorage = removeStorage(storage); // remove existing storage record if exists
+        //if (oldStorage)
+        //    storage->addWritedSpace(oldStorage->getWritedSpace());
+        m_storageRoots.insert(storage->getIndex(), storage);
+        if (storage->isStorageAvailable())
+            storage->setStatus(Qn::Online);
 
 
-    QSet<int> depracateStorageIndexes = getDeprecateIndexList(storage->getPath());
-    foreach(const int& value, depracateStorageIndexes)
-        m_storageRoots.insert(value, storage);
+        QSet<int> depracateStorageIndexes = getDeprecateIndexList(storage->getPath());
+        foreach(const int& value, depracateStorageIndexes)
+            m_storageRoots.insert(value, storage);
 
-    connect(storage.data(), SIGNAL(archiveRangeChanged(const QnAbstractStorageResourcePtr &, qint64, qint64)), 
-            this, SLOT(at_archiveRangeChanged(const QnAbstractStorageResourcePtr &, qint64, qint64)), Qt::DirectConnection);
+        connect(storage.data(), SIGNAL(archiveRangeChanged(const QnAbstractStorageResourcePtr &, qint64, qint64)), 
+                this, SLOT(at_archiveRangeChanged(const QnAbstractStorageResourcePtr &, qint64, qint64)), Qt::DirectConnection);
+    }
     loadFullFileCatalog(storage);
 }
 
@@ -530,8 +592,13 @@ void QnStorageManager::clearSpace()
     if (!m_catalogLoaded)
         return;
 
-    foreach(const QnStorageDbPtr& sdb, m_chunksDB)
-        sdb->beforeDelete();
+    {
+        QMutexLocker lock(&m_sdbMutex);
+        foreach(const QnStorageDbPtr& sdb, m_chunksDB) {
+            if (sdb)
+                sdb->beforeDelete();
+        }
+    }
 
     // 1. delete old data if cameras have max duration limit
     clearMaxDaysData();
@@ -543,8 +610,13 @@ void QnStorageManager::clearSpace()
     foreach(const QnStorageResourcePtr& storage, storages)
         clearOldestSpace(storage, false);
 
-    foreach(const QnStorageDbPtr& sdb, m_chunksDB)
-        sdb->afterDelete();
+    {
+        QMutexLocker lock(&m_sdbMutex);
+        foreach(const QnStorageDbPtr& sdb, m_chunksDB) {
+            if (sdb)
+                sdb->afterDelete();
+        }
+    }
 
     clearUnusedMotion();
     //clearCameraHistory();
@@ -577,8 +649,9 @@ void QnStorageManager::clearDbByChunk(DeviceFileCatalogPtr catalog, const Device
     {
         QnStorageResourcePtr storage = storageRoot(chunk.storageIndex);
         if (storage) {
-            QnStorageDbPtr sdb = m_chunksDB[storage->getPath()];
-            sdb->deleteRecords(catalog->cameraUniqueId(), catalog->getRole(), chunk.startTimeMs);
+            QnStorageDbPtr sdb = getSDB(storage);
+            if (sdb)
+                sdb->deleteRecords(catalog->cameraUniqueId(), catalog->getRole(), chunk.startTimeMs);
         }
     }
 }
@@ -829,6 +902,12 @@ void QnStorageManager::stopAsyncTasks()
         delete m_asyncRebuildTask;
         m_asyncRebuildTask = 0;
     }
+
+    if (m_asyncPartialScan) {
+        m_asyncPartialScan->stop();
+        delete m_asyncPartialScan;
+        m_asyncPartialScan = 0;
+    }
 }
 
 void QnStorageManager::updateStorageStatistics()
@@ -1017,8 +1096,9 @@ void QnStorageManager::replaceChunks(const QnTimePeriod& rebuildPeriod, const Qn
     if (recordingTime > 0)
         ownCatalog->setLatRecordingTime(recordingTime);
 
-    QnStorageDbPtr sdb = m_chunksDB[storage->getPath()];
-    sdb->replaceChunks(cameraUniqueId, catalog, newCatalog->m_chunks);
+    QnStorageDbPtr sdb = getSDB(storage);
+    if (sdb)
+        sdb->replaceChunks(cameraUniqueId, catalog, newCatalog->m_chunks);
 }
 
 DeviceFileCatalogPtr QnStorageManager::getFileCatalogInternal(const QString& cameraUniqueId, QnServer::ChunksCatalog catalog)
@@ -1089,8 +1169,9 @@ bool QnStorageManager::fileFinished(int durationMs, const QString& fileName, QnA
     DeviceFileCatalogPtr catalog = getFileCatalog(cameraUniqueId, quality);
     if (catalog == 0)
         return false;
-    QnStorageDbPtr sdb = m_chunksDB[storage->getPath()];
-    sdb->addRecord(cameraUniqueId, DeviceFileCatalog::catalogByPrefix(quality), catalog->updateDuration(durationMs, fileSize));
+    QnStorageDbPtr sdb = getSDB(storage);
+    if (sdb)
+        sdb->addRecord(cameraUniqueId, DeviceFileCatalog::catalogByPrefix(quality), catalog->updateDuration(durationMs, fileSize));
     return true;
 }
 
@@ -1152,12 +1233,16 @@ void QnStorageManager::doMigrateCSVCatalog(QnServer::ChunksCatalog catalog)
             {
                 QnStorageResourcePtr storage = findStorageByOldIndex(chunk.storageIndex, storageIndexes);
                 if (storage) {
-                    QnStorageDbPtr sdb = m_chunksDB[storage->getPath()];
-                    sdb->addRecord(mac, catalog, chunk);
+                    QnStorageDbPtr sdb = getSDB(storage);
+                    if (sdb)
+                        sdb->addRecord(mac, catalog, chunk);
                 }
             }
-            foreach(const QnStorageDbPtr& sdb, m_chunksDB.values())
-                sdb->flushRecords();
+            QMutexLocker lock(&m_sdbMutex);
+            foreach(const QnStorageDbPtr& sdb, m_chunksDB.values()) {
+                if (sdb)
+                    sdb->flushRecords();
+            }
             QFile::remove(catalogName);
             QDir dir;
         }
@@ -1205,7 +1290,7 @@ bool QnStorageManager::addBookmark(const QByteArray &cameraGuid, QnCameraBookmar
     if (!storage)
         return false;
 
-    QnStorageDbPtr sdb = m_chunksDB[storage->getPath()];
+    QnStorageDbPtr sdb = getSDB(storage);
     if (!sdb)
         return false;
 
@@ -1228,7 +1313,7 @@ bool QnStorageManager::updateBookmark(const QByteArray &cameraGuid, QnCameraBook
     if (!storage)
         return false;
 
-    QnStorageDbPtr sdb = m_chunksDB[storage->getPath()];
+    QnStorageDbPtr sdb = getSDB(storage);
     if (!sdb)
         return false;
 
@@ -1255,7 +1340,7 @@ bool QnStorageManager::deleteBookmark(const QByteArray &cameraGuid, QnCameraBook
     if (!storage)
         return false;
 
-    QnStorageDbPtr sdb = m_chunksDB[storage->getPath()];
+    QnStorageDbPtr sdb = getSDB(storage);
     if (!sdb)
         return false;
 
@@ -1266,7 +1351,9 @@ bool QnStorageManager::deleteBookmark(const QByteArray &cameraGuid, QnCameraBook
 }
 
 
-bool QnStorageManager::getBookmarks(const QByteArray &cameraGuid, const QnCameraBookmarkSearchFilter &filter, QnCameraBookmarkList &result) {
+bool QnStorageManager::getBookmarks(const QByteArray &cameraGuid, const QnCameraBookmarkSearchFilter &filter, QnCameraBookmarkList &result) 
+{
+    QMutexLocker lock(&m_sdbMutex);
     foreach (const QnStorageDbPtr &sdb, m_chunksDB) {
         if (!sdb)
             continue;
